Prepares to deal with topics as a component in the main GameObject.
In order to manage messages exchange by mean a topic mecanism, It seems that, in Unity phylosophy, its is difficult to call GameObjects methods from other defined classes. Those kinds of methods call have to be from a monobehaviour script (a script as a component of a GameObject). So,  a possible way to overcome this would be by considering the script defining the topics as associated scripts to the main GameObject (As script components).