First version: CubeCollector Game.
This is the first version of the CubeCollector game, where an agent manage the player movements (randomly) and tries to collect the needed cubes to win. After having the total collected cubes reach the needed number (the latter is indicated by Gama agent at the beginning by mean of notification subscribe message), a notification message from unity to Gama agent is sent, thus, ending the game.